[ refactor ] NameEmail 한글 갈아끼우기
