+++ src/core/signup/buttonText.ts
